Update get_loop logic, and move it to utils.common
