+++ tests/utils/common.py
@@ -1,4 +1,7 @@
-from biothings.utils.common import merge
+from concurrent.futures import ThreadPoolExecutor
+
+from biothings.utils.common import get_loop_with_max_workers, merge
+
 
 def test_merge_0():
     x = {}
@@ -57,6 +60,18 @@ def test_merge_5():
     print(x)
 
 
+def test_get_loop():
+    # Given
+    max_workers = 2
+
+    # Action
+    loop = get_loop_with_max_workers(max_workers=max_workers)
+
+    # Asserts
+    assert isinstance(loop._default_executor, ThreadPoolExecutor)
+    assert loop._default_executor._max_workers == max_workers
+
+
 if __name__ == '__main__':
     test_merge_0()
     test_merge_1()
@@ -64,6 +79,8 @@ if __name__ == '__main__':
     test_merge_3()
     test_merge_4()
     test_merge_5()
+    test_get_loop()
+
 
 # ANS
 # {}
